# application/code/scheduling-manager/swagger/swagger.json
{
   "openapi": "3.0.1",
   "info": {
      "title": "BBrown Capstone Project - Hangout Point",
      "description": "This is a RESTful API for managing bookings for BookMyShow.",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
         "email": "bbrown.caltech@gmail.com"
      },
      "license": {
         "name": "MIT",
         "url": "https://opensource.org/licenses/MIT"
      },
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "http://localhost:3000"
      }
   ],
   "tags": [
      {
         "name": "schedule",
         "description": "Manage scheduled bookings"
      }
   ],
   "paths": {
      "/schedule": {
         "post": {
            "tags": [
               "schedule"
            ],
            "summary": "Add a new scheduled booking record",
            "operationId": "addBooking",
            "requestBody": {
               "description": "Scheduled booking that needs to be added",
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/booking"
                     }
                  },
                  "application/xml": {
                     "schema": {
                        "$ref": "#/components/schemas/booking"
                     }
                  }
               },
               "required": true
            },
            "responses": {
               "201": {
                  "description": "Item created",
                  "content": {
                     "application/xml": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     },
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     }
                  }
               },
               "405": {
                  "description": "Invalid input",
                  "content": {}
               }
            },
            "x-codegen-request-body-name": "body"
         },
         "get": {
            "tags": [
               "schedule"
            ],
            "summary": "Return all scheduled bookings",
            "operationId": "getBooking",
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/xml": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/booking"
                           }
                        }
                     },
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "$ref": "#/components/schemas/booking"
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/schedule/{bookingId}": {
         "put": {
            "tags": [
               "schedule"
            ],
            "summary": "Update an existing booking record",
            "operationId": "updateBooking",
            "parameters": [
               {
                  "name": "bookingId",
                  "in": "path",
                  "description": "Booking id to update",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "requestBody": {
               "description": "Scheduled booking that needs to be updated",
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/booking"
                     }
                  },
                  "application/xml": {
                     "schema": {
                        "$ref": "#/components/schemas/booking"
                     }
                  }
               },
               "required": true
            },
            "responses": {
               "200": {
                  "description": "Scheduled booking updated",
                  "content": {
                     "application/xml": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     },
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied",
                  "content": {}
               },
               "404": {
                  "description": "Scheduled booking not found",
                  "content": {}
               },
               "405": {
                  "description": "Validation exception",
                  "content": {}
               }
            },
            "x-codegen-request-body-name": "body"
         },
         "get": {
            "tags": [
               "schedule"
            ],
            "summary": "Find scheduled booking by ID",
            "description": "Returns a single sheduled booking",
            "operationId": "getBookingId",
            "parameters": [
               {
                  "name": "bookingId",
                  "in": "path",
                  "description": "ID of scheduled booking to return",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/xml": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     },
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/booking"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied",
                  "content": {}
               },
               "404": {
                  "description": "Scheduled booking not found",
                  "content": {}
               }
            }
         },
         "delete": {
            "tags": [
               "schedule"
            ],
            "summary": "Deletes a booking item",
            "operationId": "deleteBooking",
            "parameters": [
               {
                  "name": "bookingId",
                  "in": "path",
                  "description": "Booking id to delete",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "responses": {
               "400": {
                  "description": "Invalid ID supplied",
                  "content": {}
               },
               "404": {
                  "description": "Scheduled booking not found",
                  "content": {}
               }
            }
         }
      }
   },
   "components": {
      "schemas": {
         "booking": {
            "type": "object",
            "properties": {
               "_id": {
                  "type": "string"
               },
               "customerName": {
                  "type": "string"
               },
               "showName": {
                  "type": "string"
               },
               "showDate": {
                  "type": "string",
                  "format": "date"
               },
               "totalTickets": {
                  "type": "integer",
                  "format": "int32"
               }
            },
            "xml": {
               "name": "Booking"
            }
         }
      }
   }
}